requirements.txt
piexif==1.1.3
pyyaml==6.0.2